Refactor tests
Moves all tests for UpdatingElement API to `updating-element_test.ts` from `lit-element_test.ts`.
